fix: started to reduce cyclic dependencies in code
