+++ src/query-builder/query.ts
@@ -1,33 +1,109 @@
+import { TermJson, ComplexTermJson } from '../internal-types';
+import { TermConfig } from './query-config';
+import { RQuery, RethinkDBErrorType } from '../types';
 import { RethinkDBError } from '../error/error';
-import { backtraceTerm } from '../error/term-backtrace';
-import { TermJson } from '../internal-types';
-import { bracket, termConfig } from './query-config';
-import {
-  doTermFunc,
-  getCursorQueryFunc,
-  runQueryFunc,
-  termBuilder,
-} from './term-builder';
+import { TermType } from '../proto/enums';
+import { parseParam } from './param-parser';
+import { parseOptarg } from './parse-opt-arg';
 
 export const querySymbol = Symbol('RethinkDBQuery');
 
-export const isQuery = (query: any) =>
+export type RunnableRQuery = RQuery & {
+  term: TermJson;
+  [querySymbol]: true;
+  run: () => any;
+  getCursor: () => any;
+  do: () => any;
+};
+
+export const isQuery = (query: any): query is RunnableRQuery =>
   query === Object(query) && Object.hasOwnProperty.call(query, querySymbol);
 
-export function toQuery(term: TermJson) {
-  const query: any = termBuilder(bracket, term);
-  query.term = term;
+const numToStringArr = ['', 'First', 'Second', 'Third', 'Fourth', 'Fifth'];
+function numToString(num: number) {
+  return numToStringArr.map((_, i) => i).includes(num)
+    ? numToStringArr[num]
+    : num.toString();
+}
 
-  query[querySymbol] = true;
-
-  query.toString = () => backtraceTerm(term)[0];
-  query.run = runQueryFunc(term);
-  query.getCursor = getCursorQueryFunc(term);
-  query.do = doTermFunc(query);
-
-  for (let i = 0; i < termConfig.length; i += 1) {
-    const config = termConfig[i];
-    query[config[1]] = termBuilder(config, term);
-  }
-  return query;
+export function termBuilder(
+  [termType, termName, minArgs, maxArgs, optargType]: TermConfig,
+  toQuery: (term: TermJson) => RunnableRQuery,
+  currentTerm?: TermJson,
+) {
+  return (...args: any[]) => {
+    let optarg: Record<string, unknown> | undefined;
+    const params: TermJson[] = currentTerm !== undefined ? [currentTerm] : [];
+    if (isQuery(args[0]) && args[0].term[0] === TermType.ARGS) {
+      params.push(parseParam(args[0]));
+      optarg = optargType !== false ? args[1] : undefined;
+    } else {
+      const argsLength = args.length;
+      if (minArgs === maxArgs && argsLength !== minArgs) {
+        throw new RethinkDBError(
+          `\`${
+            !currentTerm ? `r.${termName}` : termName
+          }\` takes ${minArgs} argument${
+            minArgs === 1 ? '' : 's'
+          }, ${argsLength} provided${!currentTerm ? '.' : ' after:'}`,
+          { term: currentTerm, type: RethinkDBErrorType.ARITY },
+        );
+      }
+      if (argsLength < minArgs) {
+        throw new RethinkDBError(
+          `\`${
+            !currentTerm ? `r.${termName}` : termName
+          }\` takes at least ${minArgs} argument${
+            minArgs === 1 ? '' : 's'
+          }, ${argsLength} provided${!currentTerm ? '.' : ' after:'}`,
+          { term: currentTerm, type: RethinkDBErrorType.ARITY },
+        );
+      }
+      if (maxArgs !== -1 && argsLength > maxArgs) {
+        throw new RethinkDBError(
+          `\`${
+            !currentTerm ? `r.${termName}` : termName
+          }\` takes at most ${maxArgs} argument${
+            maxArgs === 1 ? '' : 's'
+          }, ${argsLength} provided${!currentTerm ? '.' : ' after:'}`,
+          { term: currentTerm, type: RethinkDBErrorType.ARITY },
+        );
+      }
+      switch (optargType) {
+        case 'last':
+          optarg = parseOptarg(args[maxArgs - 1]);
+          break;
+        case 'required':
+        case 'optional':
+        case 'last-optional':
+          optarg = parseOptarg(args[argsLength - 1]);
+      }
+      if (
+        !optarg &&
+        (optargType === 'required' ||
+          (argsLength === maxArgs &&
+            ['last', 'last-optional'].includes(optargType as any)))
+      ) {
+        throw new RethinkDBError(
+          `${numToString(
+            argsLength,
+          )} argument of \`${termName}\` must be an object.`,
+          { term: currentTerm, type: RethinkDBErrorType.ARITY },
+        );
+      }
+      params.push(
+        ...args
+          .filter((_, i) => (optarg ? i < argsLength - 1 : true))
+          .map((x) => parseParam(x)),
+      );
+    }
+    const term: ComplexTermJson = [termType];
+    if (params.length > 0) {
+      term[1] = params;
+    }
+    if (optarg) {
+      term[2] = optarg;
+    }
+    return toQuery(term);
+  };
 }
