+++ src/query-builder/r.ts
@@ -1,7 +1,6 @@
 import { RethinkDBConnection } from '../connection/connection';
 import { MasterConnectionPool } from '../connection/master-pool';
 import { RethinkDBError } from '../error/error';
-import { TermType } from '../proto/enums';
 import {
   R,
   RConnectionOptions,
@@ -10,9 +9,18 @@ import {
   RQuery,
 } from '../types';
 import { globals } from './globals';
-import { toQuery } from './query';
 import { funcall, rConfig, rConsts, termConfig } from './query-config';
-import { expr, termBuilder } from './term-builder';
+import { isQuery, termBuilder } from './query';
+import { toQuery } from './query-runner';
+import { validateTerm } from './validate-term';
+import { parseParam } from './param-parser';
+
+const expr = (arg: any, nestingLevel: number = globals.nestingLevel) => {
+  if (isQuery(arg)) {
+    return arg;
+  }
+  return toQuery(parseParam(arg, nestingLevel));
+};
 
 export const r: R = expr as any;
 r.connectPool = async (options: RPoolConnectionOptions = {}) => {
@@ -68,7 +76,7 @@ r.connect = async (options: RConnectionOptions = {}) => {
       { type: RethinkDBErrorType.API_FAIL },
     );
   }
-  const c = new RethinkDBConnection(server, options as any);
+  const c = new RethinkDBConnection(server, options);
   await c.reconnect();
   return c;
 };
@@ -81,63 +89,35 @@ r.waitForHealthy = () => {
     type: RethinkDBErrorType.MASTER_POOL_FAIL,
   });
 };
-r.setNestingLevel = (level: number) => (globals.nestingLevel = level);
-r.setArrayLimit = (limit?: number) => (globals.arrayLimit = limit);
-r.serialize = (termStr: RQuery) => JSON.stringify((termStr as any).term);
+r.setNestingLevel = (level: number) => {
+  globals.nestingLevel = level;
+};
+r.setArrayLimit = (limit?: number) => {
+  globals.arrayLimit = limit;
+};
+r.serialize = (termStr: RQuery) => JSON.stringify(termStr.term);
 r.deserialize = (termStr: string) => toQuery(validateTerm(JSON.parse(termStr)));
 r.expr = expr;
 r.do = (...args: any[]) => {
   const last = args.pop();
-  return termBuilder(funcall)(last, ...args);
+  return termBuilder(funcall, toQuery)(last, ...args);
 };
-rConfig.forEach((config) => ((r as any)[config[1]] = termBuilder(config)));
+rConfig.forEach(
+  (config) => ((r as any)[config[1]] = termBuilder(config, toQuery)),
+);
 rConsts.forEach(([type, name]) => ((r as any)[name] = toQuery([type])));
 termConfig
   .filter(([_, name]) => !(name in r))
   .forEach(
     ([type, name, minArgs, maxArgs, optArgs]) =>
-      ((r as any)[name] = termBuilder([
-        type,
-        name,
-        minArgs + 1,
-        maxArgs === -1 ? maxArgs : maxArgs + 1,
-        optArgs,
-      ])),
+      ((r as any)[name] = termBuilder(
+        [
+          type,
+          name,
+          minArgs + 1,
+          maxArgs === -1 ? maxArgs : maxArgs + 1,
+          optArgs,
+        ],
+        toQuery,
+      )),
   );
-
-function validateTerm(term: any): any {
-  if (term === undefined) {
-    throw new RethinkDBError(`Invalid term:\n${JSON.stringify(term)}\n`);
-  }
-  if (typeof term === 'function') {
-    throw new RethinkDBError(`Invalid term:\n${term.toString()}\n`);
-  }
-  if (typeof term === 'object') {
-    if (Array.isArray(term)) {
-      if (term.length > 3) {
-        throw new RethinkDBError(`Invalid term:\n${JSON.stringify(term)}\n`);
-      }
-      const [func, args, options] = term;
-      if (typeof func !== 'number' || TermType[func] === undefined) {
-        throw new RethinkDBError(`Invalid term:\n${JSON.stringify(term)}\n`);
-      }
-      if (args !== undefined) {
-        if (!Array.isArray(args)) {
-          throw new RethinkDBError(`Invalid term:\n${JSON.stringify(term)}\n`);
-        }
-        if (!args.every((arg) => validateTerm(arg))) {
-          throw new RethinkDBError(`Invalid term:\n${JSON.stringify(term)}\n`);
-        }
-      }
-      if (
-          options !== undefined &&
-        !Object.values(term).every((value) => validateTerm(value))
-      ) {
-        throw new RethinkDBError(`Invalid term:\n${JSON.stringify(term)}\n`);
-      }
-    } else if (!Object.values(term).every((value) => validateTerm(value))) {
-      throw new RethinkDBError(`Invalid term:\n${JSON.stringify(term)}\n`);
-    }
-  }
-  return term;
-}
